requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
pymongo==4.5.0
mysql-connector-python==8.1.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
joblib==1.3.2
Werkzeug==2.3.7
python-dotenv==1.0.0
redis==5.0.0
celery==5.3.4
gunicorn==21.2.0
email-validator==2.0.0